pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-agents"
version = "0.1.0"
description = "Master Project"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic-settings>=2.2",
    "celery>=5.6.2"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]